订单统计业务代码

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -67,5 +67,5 @@ public interface OrderMapper {
 
     Double subByMap(Map map);
 
-    Double countByMap(Map map);
+    Integer countByMap(Map map);
 }
